attendance-reader/src/requirements.txt
boto3==1.37.26
python-dotenv==1.1.0
nfcpy==1.0.4
fastapi==0.115.12
uvicorn==0.34.0